+++ lib/strong_migrations/checker.rb
@@ -130,7 +130,7 @@ Then add the NOT NULL constraint in separate migrations."
                 # safe to change varchar to text (and text to text)
                 safe = ["character varying", "text", "citext"].include?(existing_type)
               when "citext"
-                safe = ["text"].include?(existing_type)
+                safe = ["text", "character varying"].include?(existing_type)
               when "numeric", "decimal"
                 # numeric and decimal are equivalent and can be used interchangably
                 safe = ["numeric", "decimal"].include?(existing_type) &&