+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.8.1 (unreleased)
 
 - Changing between `text` and `citext` is safe in Postgres
+- Changing a `varchar` column to a `citext` column is safe in Postgres
 
 ## 0.8.0 (2022-02-09)